WifiStateMachine: force autoconnect at bootup
Otherwise during bootup the following case happens:
- SupplicantState is SCANNING
- getLastSelectedConfiguration() is null
- mConnectionRequests is 0

In that case, attemptAutoJoin is set to false which causes the
system not to connect although once the scanning is done some
known networks are discovered.

A better solution is most likely possible but not sure where to
start. It can be the scanning that is too slow, some delay is
required or a proper call to increase mConnectionRequests.

Change-Id: I7cd76845e7f624eab1e354b54000a323e33be373
Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>